User Controller Refactor and fixes
