+++ src/controllers/UserController.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
-import { prisma } from '../db';
 import { CreateUser, LoginUser, RequestWithAuth } from "../ts/index";
 import { RequestStatuses } from "../ts/enums";
 import UserService from "../services/UserService";
-import ForbiddenError from "../errors/ForbiddenError";
+import ForbiddenError from "../errors/clienterrors/ForbiddenError";
+import User from "../models/User";
 
 export default class UserController {
     static async createUser(req: Request, res: Response): Promise<Object> {
@@ -31,46 +31,12 @@ export default class UserController {
             }
         });
     }
-    // FIXME: Remove this debug method
-    static async getUsers(req: Request, res: Response): Promise<Object> {
-        const users = await prisma.user.findMany({
-            select: {
-                login: true,
-                name: true,
-                role: {
-                    select: {
-                        name: true,
-                        permissions: {
-                            select: {
-                                name: true,
-                            }
-                        }
-                    }
-                }
-            }
-        });
-        return res.json({
-            users
-        });
-    }
 
     static async profile(req: RequestWithAuth, res: Response): Promise<Object> {
         // FIXME: Refactor to middleware
         const { id } = req.auth!;
-        const user = await prisma.user.findFirst({
-            where: {
-                id
-            },
-            include: {
-                role: {
-                    include: {
-                        permissions: true
-                    }
-                }
-            },
-        });
+        const user = await User.findUserByIdWithRolePermission(id)
         if (!user) throw new ForbiddenError('Unauthorized');
-        // END
 
         return res.status(RequestStatuses.OK).json({
             data: {
