+++ src/routes/UserRouter.ts
@@ -5,7 +5,6 @@ import { registrationValidation, loginValidation } from "../validators/UserValid
 import { wrap } from '../middleware/errorHandler';
 const router = Router();
 
-router.get("/", wrap(UserController.getUsers));
 router.post("/register", registrationValidation(), wrap(UserController.createUser));
 router.post("/login", loginValidation(), wrap(UserController.loginUser));
 router.get("/profile", auth, wrap(UserController.profile));
