+++ src/services/UserService.ts
@@ -1,5 +1,4 @@
-
-import UnauthorizedError from "../errors/UnauthorizedError";
+import UnauthorizedError from "../errors/clienterrors/UnauthorizedError";
 import { Encrypt } from "../helper/encrypt";
 import RefreshToken from "../models/RefreshToken";
 import User from "../models/User";
@@ -20,14 +19,11 @@ export default class UserService {
 
     public static async login(userData: LoginUser) {
         const {login, password} = userData;
-
         const user = await User.findUserByLogin(login);
         
         if (!user) throw new UnauthorizedError("Login or password incorrect");
         if (!await Encrypt.comparePassword(password, user.password)) throw new UnauthorizedError("Login or password incorrect");
-        
         const { id } = user;
-
         const token = jwt.sign({ id }, process.env.tokenSecret!, { expiresIn: process.env.tokenLife })
         const refreshToken = jwt.sign({ id }, process.env.tokenRefreshSecret!, { expiresIn: process.env.tokenRefreshLife })
